Added location transforming for brand and series in simple model
Refactored some related code from the ItemModelTransformer
into separate classes to prevent code duplication.
