+++ pom.xml
@@ -6,7 +6,7 @@
     <groupId>org.atlasapi</groupId>
     <artifactId>atlas</artifactId>
     <packaging>war</packaging>
-    <version>5.0-SNAPSHOT</version>
+    <version>5.1-SNAPSHOT</version>
     <description>Atlas 3.0 API</description>
     <build>
         <finalName>atlas</finalName>
@@ -421,7 +421,7 @@
         <dependency>
             <groupId>org.atlasapi</groupId>
             <artifactId>atlas-model</artifactId>
-            <version>${atlas.version}</version>
+            <version>5.1-SNAPSHOT</version>
             <exclusions>
                 <exclusion>
                     <groupId>com.google.guava</groupId>
@@ -843,7 +843,7 @@
 
     </dependencies>
     <properties>
-        <atlas.version>${project.version}</atlas.version>
+        <atlas.version>5.0-SNAPSHOT</atlas.version>
         <common.version>1.0-SNAPSHOT</common.version>
         <netbeans.hint.deploy.server>Tomcat60</netbeans.hint.deploy.server>
         <jetty.version>9.4.32.v20200930</jetty.version>
