+++ src/main/java/org/atlasapi/input/ItemModelTransformer.java
@@ -1,23 +1,20 @@
 package org.atlasapi.input;
 
-import java.util.Currency;
-import java.util.Date;
-import java.util.Set;
-
-import org.atlasapi.media.TransportSubType;
-import org.atlasapi.media.TransportType;
+import com.google.common.base.Optional;
+import com.google.common.collect.ImmutableSet;
+import com.google.common.collect.ImmutableSet.Builder;
+import com.google.common.collect.Iterables;
+import com.google.common.collect.Sets;
+import com.metabroadcast.common.ids.NumberToShortStringCodec;
+import com.metabroadcast.common.intl.Countries;
+import com.metabroadcast.common.time.Clock;
 import org.atlasapi.media.channel.ChannelResolver;
 import org.atlasapi.media.entity.Broadcast;
 import org.atlasapi.media.entity.Encoding;
 import org.atlasapi.media.entity.Episode;
 import org.atlasapi.media.entity.Film;
 import org.atlasapi.media.entity.Item;
-import org.atlasapi.media.entity.Location;
 import org.atlasapi.media.entity.ParentRef;
-import org.atlasapi.media.entity.Policy;
-import org.atlasapi.media.entity.Policy.Platform;
-import org.atlasapi.media.entity.Policy.RevenueContract;
-import org.atlasapi.media.entity.Provider;
 import org.atlasapi.media.entity.ReleaseDate;
 import org.atlasapi.media.entity.Restriction;
 import org.atlasapi.media.entity.Song;
@@ -25,28 +22,20 @@ import org.atlasapi.media.entity.Version;
 import org.atlasapi.media.segment.SegmentEvent;
 import org.atlasapi.persistence.lookup.entry.LookupEntryStore;
 import org.atlasapi.persistence.topic.TopicStore;
-
-import com.metabroadcast.common.currency.Price;
-import com.metabroadcast.common.ids.NumberToShortStringCodec;
-import com.metabroadcast.common.intl.Countries;
-import com.metabroadcast.common.media.MimeType;
-import com.metabroadcast.common.time.Clock;
-import com.metabroadcast.common.time.DateTimeZones;
-
-import com.google.common.base.Optional;
-import com.google.common.collect.ImmutableSet;
-import com.google.common.collect.ImmutableSet.Builder;
-import com.google.common.collect.Iterables;
-import com.google.common.collect.Sets;
 import org.joda.time.DateTime;
 import org.joda.time.Duration;
 import org.joda.time.LocalDate;
+
+import java.util.Date;
+import java.util.Set;
 
 import static com.google.common.base.Preconditions.checkNotNull;
 
 public class ItemModelTransformer extends ContentModelTransformer<org.atlasapi.media.entity.simple.Item, Item> {
 
     private final BroadcastModelTransformer broadcastTransformer;
+    private final EncodingModelTransformer encodingTransformer;
+    private final RestrictionModelTransformer restrictionTransformer;
     private final SegmentModelTransformer segmentModelTransformer;
 
     public ItemModelTransformer(LookupEntryStore lookupStore, TopicStore topicStore,
@@ -54,6 +43,8 @@ public class ItemModelTransformer extends ContentModelTransformer<org.atlasapi.m
             ClipModelTransformer clipsModelTransformer, Clock clock, SegmentModelTransformer segmentModelTransformer) {
         super(lookupStore, topicStore, idCodec, clipsModelTransformer, clock);
         this.broadcastTransformer = BroadcastModelTransformer.create(channelResolver);
+        this.encodingTransformer = EncodingModelTransformer.create();
+        this.restrictionTransformer = RestrictionModelTransformer.create();
         this.segmentModelTransformer = checkNotNull(segmentModelTransformer);
     }
 
@@ -200,18 +191,13 @@ public class ItemModelTransformer extends ContentModelTransformer<org.atlasapi.m
     }
 
     private Optional<Restriction> createRestriction(
-            org.atlasapi.media.entity.simple.Restriction simpleRestriction) {
+            org.atlasapi.media.entity.simple.Restriction simpleRestriction
+    ) {
         if (simpleRestriction == null) {
             return Optional.absent();
         }
-        
-        Restriction restriction = new Restriction();
-        
-        restriction.setRestricted(simpleRestriction.isRestricted());
-        restriction.setAuthority(simpleRestriction.getAuthority());
-        restriction.setRating(simpleRestriction.getRating());
-        restriction.setMinimumAge(simpleRestriction.getMinimumAge());
-        restriction.setMessage(simpleRestriction.getMessage());
+
+        Restriction restriction = restrictionTransformer.transform(simpleRestriction);
 
         return Optional.of(restriction);
     }
@@ -219,95 +205,10 @@ public class ItemModelTransformer extends ContentModelTransformer<org.atlasapi.m
     private Set<Encoding> encodingsFrom(Set<org.atlasapi.media.entity.simple.Location> locations, DateTime now) {
         Builder<Encoding> encodings = ImmutableSet.builder();
         for (org.atlasapi.media.entity.simple.Location simpleLocation : locations) {
-            Encoding encoding = encodingFrom(simpleLocation, now);
-            Location location = locationFrom(simpleLocation, now);
-            Policy policy = policyFrom(simpleLocation, now);
-            location.setPolicy(policy);
-            encoding.setAvailableAt(ImmutableSet.of(location));
+            Encoding encoding = encodingTransformer.transform(simpleLocation, now);
             encodings.add(encoding);
         }
         return encodings.build();
-    }
-
-    private Encoding encodingFrom(org.atlasapi.media.entity.simple.Location inputLocation, DateTime now) {
-        Encoding encoding = new Encoding();
-        encoding.setLastUpdated(now);
-        encoding.setAdvertisingDuration(inputLocation.getAdvertisingDuration());
-        encoding.setAudioBitRate(inputLocation.getAudioBitRate());
-        encoding.setAudioChannels(inputLocation.getAudioChannels());
-        encoding.setAudioCoding(asMimeType(inputLocation.getAudioCoding()));
-        encoding.setBitRate(inputLocation.getBitRate());
-        encoding.setContainsAdvertising(inputLocation.getContainsAdvertising());
-        encoding.setDataContainerFormat(asMimeType(inputLocation.getDataContainerFormat()));
-        encoding.setDataSize(inputLocation.getDataSize());
-        encoding.setDistributor(inputLocation.getDistributor());
-        encoding.setHasDOG(inputLocation.getHasDOG());
-        encoding.setSource(inputLocation.getSource());
-        encoding.setVideoAspectRatio(inputLocation.getVideoAspectRatio());
-        encoding.setVideoBitRate(inputLocation.getVideoBitRate());
-        encoding.setVideoCoding(asMimeType(inputLocation.getVideoCoding()));
-        encoding.setVideoFrameRate(inputLocation.getVideoFrameRate());
-        encoding.setVideoHorizontalSize(inputLocation.getVideoHorizontalSize());
-        encoding.setVideoProgressiveScan(inputLocation.getVideoProgressiveScan());
-        encoding.setVideoVerticalSize(inputLocation.getVideoVerticalSize());
-        encoding.setHighDefinition(inputLocation.getHighDefinition());
-        return encoding;
-    }
-
-    private Location locationFrom(org.atlasapi.media.entity.simple.Location inputLocation, DateTime now) {
-        Location location = new Location();
-        location.setLastUpdated(now);
-        location.setCanonicalUri(inputLocation.getCanonicalUri());
-        location.setEmbedCode(inputLocation.getEmbedCode());
-        location.setEmbedId(inputLocation.getEmbedId());
-        location.setTransportIsLive(inputLocation.getTransportIsLive());
-        location.setUri(inputLocation.getUri());
-        if(inputLocation.getProvider() != null) {
-            location.setProvider(providerFrom(inputLocation.getProvider()));
-        }
-        if (inputLocation.getTransportSubType() != null) {
-            location.setTransportSubType(TransportSubType.fromString(inputLocation.getTransportSubType()));
-        }
-        if (inputLocation.getTransportType() != null) {
-            location.setTransportType(TransportType.fromString(inputLocation.getTransportType()));
-        }
-        return location;
-    }
-
-    private Provider providerFrom(org.atlasapi.media.entity.simple.Provider simpleProvider) {
-        return new Provider(simpleProvider.getName(), simpleProvider.getIconUrl());
-    }
-
-    private Policy policyFrom(org.atlasapi.media.entity.simple.Location inputLocation, DateTime now) {
-        Policy policy = new Policy();
-        policy.setLastUpdated(now);
-        policy.setAvailabilityStart(asUtcDateTime(inputLocation.getAvailabilityStart()));
-        policy.setAvailabilityEnd(asUtcDateTime(inputLocation.getAvailabilityEnd()));
-        policy.setDrmPlayableFrom(asUtcDateTime(inputLocation.getDrmPlayableFrom()));
-        policy.setPlatform(Platform.fromKey(inputLocation.getPlatform()));
-        if (inputLocation.getCurrency() != null && inputLocation.getPrice() != null) {
-            Currency currency = Currency.getInstance(inputLocation.getCurrency());
-            policy.setPrice(new Price(currency, inputLocation.getPrice()));
-        }
-        if (inputLocation.getAvailableCountries() != null) {
-            policy.setAvailableCountries(Countries.fromCodes(inputLocation.getAvailableCountries()));
-        }
-        policy.setRevenueContract(RevenueContract.fromKey(inputLocation.getRevenueContract()));
-        return policy;
-    }
-
-    private MimeType asMimeType(String mimeType) {
-        if (mimeType != null) {
-            return MimeType.fromString(mimeType);
-        }
-        return null;
-    }
-
-    private DateTime asUtcDateTime(Date date) {
-        if (date == null) {
-            return null;
-        }
-        return new DateTime(date).withZone(DateTimeZones.UTC);
     }
 
     private Set<ReleaseDate> releaseDatesFrom(Set<org.atlasapi.media.entity.simple.ReleaseDate> releaseDates) {
